Release 0.1.28
- Big temporary commit in the light of preparing for the new RDF compatible release (version 0.2) - we need this for the thomasmore project

+++ src/main/java/com/beligum/blocks/fs/HdfsUtils.java
@@ -10,7 +10,10 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
+import java.nio.file.FileVisitResult;
 import java.nio.file.FileVisitor;
+import java.nio.file.SimpleFileVisitor;
+import java.nio.file.attribute.BasicFileAttributes;
 import java.security.SecureRandom;
 import java.util.EnumSet;
 
@@ -40,18 +43,33 @@ public class HdfsUtils
     }
     public static void recursiveDeleteLockFiles(FileContext fs, Path path) throws IOException
     {
-        RemoteIterator<FileStatus> status = fs.listStatus(path);
-        while (status.hasNext()) {
-            FileStatus fileStatus = status.next();
-            Path lockFile = HdfsResourcePath.createLockPath(fileStatus.getPath());
-            if (fs.util().exists(lockFile)) {
-                fs.delete(lockFile, false);
+        HdfsUtils.walkFileTree(fs, path, new SimpleFileVisitor<Path>()
+        {
+            @Override
+            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException
+            {
+                return this.process(fs, file);
+            }
+            @Override
+            public FileVisitResult preVisitDirectory(Path file, BasicFileAttributes attrs) throws IOException
+            {
+                return this.process(fs, file);
             }
 
-            if (fileStatus.isDirectory()) {
-                recursiveDeleteLockFiles(fs, fileStatus.getPath());
+            private FileVisitResult process(FileContext fileContext, Path path) throws IOException
+            {
+                FileVisitResult retVal = FileVisitResult.CONTINUE;
+
+                Path lockFile = HdfsResourcePath.createLockPath(path);
+                if (lockFile!=null) {
+                    if (fs.util().exists(lockFile)) {
+                        fs.delete(lockFile, false);
+                    }
+                }
+
+                return retVal;
             }
-        }
+        });
     }
     /**
      * Basic implementation of a recursive, depth-first file visitor for HDFS paths,
